Removed the parsing from the toBatchType function and created it's own method
